First stable revision
This commit marks the first stable revision of the function. Note
that there actually seems to be a significant decrease in performance.
I attribute this to the overhead caused by thread creation, and also
the size of the critical region. I will profile and run some tests,
make a few revisions, then update again soon.